+++ js/source/spreader/wheelnav.spreader.core.js
@@ -9,17 +9,28 @@ spreader = function (wheelnav) {
 
     this.wheelnav = wheelnav;
     if (this.wheelnav.spreaderEnable) {
+
+        this.spreaderHelper = new pathHelper();
+        this.spreaderHelper.centerX = this.wheelnav.centerX;
+        this.spreaderHelper.centerY = this.wheelnav.centerY;
+        this.spreaderHelper.navItemCount = this.wheelnav.navItemCount;
+        this.spreaderHelper.navAngle = this.wheelnav.navAngle;
+        this.spreaderHelper.wheelRadius = this.wheelnav.spreaderRadius;
+        this.spreaderHelper.startAngle = this.wheelnav.spreaderStartAngle;
+        this.spreaderHelper.sliceAngle = this.wheelnav.spreaderSliceAngle;
+
         var thisWheelNav = this.wheelnav;
 
         var fontAttr = { font: '100 32px Impact, Charcoal, sans-serif' };
 
-        this.spreaderCircle = thisWheelNav.raphael.circle(thisWheelNav.centerX, thisWheelNav.centerY, thisWheelNav.spreaderRadius);
+        var spreaderPath = this.wheelnav.spreaderPathFunction(this.spreaderHelper, this.wheelnav.spreaderPathCustom);
+        this.spreaderCircle = this.wheelnav.raphael.path(spreaderPath.spreaderPathString);
         this.spreaderCircle.attr(thisWheelNav.spreaderCircleAttr);
         this.spreaderCircle.click(function () {
             thisWheelNav.spreadWheel();
         });
 
-        this.spreadOnTitle = thisWheelNav.raphael.text(thisWheelNav.centerX, thisWheelNav.centerY, "+");
+        this.spreadOnTitle = thisWheelNav.raphael.text(spreaderPath.titlePosX, spreaderPath.titlePosY, "+");
         this.spreadOnTitle.attr(fontAttr);
         this.spreadOnTitle.attr(thisWheelNav.spreaderOnAttr);
         this.spreadOnTitle.id = thisWheelNav.getSpreadOnId();
@@ -27,7 +38,7 @@ spreader = function (wheelnav) {
             thisWheelNav.spreadWheel();
         });
 
-        this.spreadOffTitle = thisWheelNav.raphael.text(thisWheelNav.centerX, thisWheelNav.centerY - 3, "–");
+        this.spreadOffTitle = thisWheelNav.raphael.text(spreaderPath.titlePosX, spreaderPath.titlePosY - 3, "–");
         this.spreadOffTitle.attr(fontAttr);
         this.spreadOffTitle.attr(thisWheelNav.spreaderOffAttr);
         this.spreadOffTitle.id = thisWheelNav.getSpreadOffId();
@@ -55,3 +66,4 @@ spreader.prototype.setVisibility = function () {
         }
     }
 };
+