+++ js/source/spreader/wheelnav.spreader.js
@@ -1,0 +1,182 @@
+﻿///#source 1 1 /js/source/spreader/wheelnav.spreader.core.js
+/* ======================================================================================= */
+/* Spreader of wheel                                                                       */
+/* ======================================================================================= */
+/* ======================================================================================= */
+/* Documentation: http://wheelnavjs.softwaretailoring.net/documentation/spreader.html      */
+/* ======================================================================================= */
+
+spreader = function (wheelnav) {
+
+    this.wheelnav = wheelnav;
+    if (this.wheelnav.spreaderEnable) {
+
+        this.spreaderHelper = new pathHelper();
+        this.spreaderHelper.centerX = this.wheelnav.centerX;
+        this.spreaderHelper.centerY = this.wheelnav.centerY;
+        this.spreaderHelper.navItemCount = this.wheelnav.navItemCount;
+        this.spreaderHelper.navAngle = this.wheelnav.navAngle;
+        this.spreaderHelper.wheelRadius = this.wheelnav.spreaderRadius;
+        this.spreaderHelper.startAngle = this.wheelnav.spreaderStartAngle;
+        this.spreaderHelper.sliceAngle = this.wheelnav.spreaderSliceAngle;
+
+        var thisWheelNav = this.wheelnav;
+
+        var fontAttr = { font: '100 32px Impact, Charcoal, sans-serif' };
+
+        var spreaderPath = this.wheelnav.spreaderPathFunction(this.spreaderHelper, this.wheelnav.spreaderPathCustom);
+        this.spreaderCircle = this.wheelnav.raphael.path(spreaderPath.spreaderPathString);
+        this.spreaderCircle.attr(thisWheelNav.spreaderCircleAttr);
+        this.spreaderCircle.click(function () {
+            thisWheelNav.spreadWheel();
+        });
+
+        this.spreadOnTitle = thisWheelNav.raphael.text(spreaderPath.titlePosX, spreaderPath.titlePosY, "+");
+        this.spreadOnTitle.attr(fontAttr);
+        this.spreadOnTitle.attr(thisWheelNav.spreaderOnAttr);
+        this.spreadOnTitle.id = thisWheelNav.getSpreadOnId();
+        this.spreadOnTitle.click(function () {
+            thisWheelNav.spreadWheel();
+        });
+
+        this.spreadOffTitle = thisWheelNav.raphael.text(spreaderPath.titlePosX, spreaderPath.titlePosY - 3, "–");
+        this.spreadOffTitle.attr(fontAttr);
+        this.spreadOffTitle.attr(thisWheelNav.spreaderOffAttr);
+        this.spreadOffTitle.id = thisWheelNav.getSpreadOffId();
+        this.spreadOffTitle.click(function () {
+            thisWheelNav.spreadWheel();
+        });
+
+        this.setVisibility();
+    }
+
+    return this;
+};
+
+spreader.prototype.setVisibility = function () {
+    if (this.wheelnav.spreaderEnable) {
+        this.spreaderCircle.toFront();
+
+        if (this.wheelnav.currentPercent > this.wheelnav.minPercent) {
+            this.spreadOffTitle.toFront();
+            this.spreadOnTitle.toBack();
+        }
+        else {
+            this.spreadOffTitle.toBack();
+            this.spreadOnTitle.toFront();
+        }
+    }
+};
+
+
+///#source 1 1 /js/source/spreader/wheelnav.spreaderPathStart.js
+/* ======================================================================================= */
+/* Spreader path definitions.                                                              */
+/* ======================================================================================= */
+
+spreaderPath = function () {
+
+    this.NullSpreader = function (helper, custom) {
+
+        helper.setBaseValue(custom);
+
+        return {
+            slicePathString: "",
+            titlePosX: helper.titlePosX,
+            titlePosY: helper.titlePosY
+        };
+    };
+
+
+
+///#source 1 1 /js/source/spreader/wheelnav.spreaderPath.Pie.js
+
+this.PieSpreaderCustomization = function () {
+
+    var custom = new spreaderPathCustomization();
+    custom.spreaderRadius = 25;
+    custom.arcBaseRadiusPercent = 1;
+    custom.arcRadiusPercent = 1;
+    custom.startRadiusPercent = 0;
+    return custom;
+};
+
+this.PieSpreader = function (helper, custom) {
+
+    if (custom === null) {
+        custom = PieSpreaderCustomization();
+    }
+
+    helper.setBaseValue(custom.spreaderPercent, custom);
+
+    var arcBaseRadius = helper.sliceRadius * custom.arcBaseRadiusPercent;
+    var arcRadius = helper.sliceRadius * custom.arcRadiusPercent;
+
+    spreaderPathString = [helper.MoveTo(helper.startAngle, arcBaseRadius),
+                 helper.ArcTo(arcRadius, helper.middleAngle, arcBaseRadius),
+                 helper.ArcTo(arcRadius, helper.endAngle, arcBaseRadius),
+                 helper.Close()];
+    
+    return {
+        spreaderPathString: spreaderPathString,
+        titlePosX: helper.titlePosX,
+        titlePosY: helper.titlePosY
+    };
+};
+
+///#source 1 1 /js/source/spreader/wheelnav.spreaderPath.Star.js
+
+this.StarSpreaderCustomization = function () {
+
+    var custom = new spreaderPathCustomization();
+    custom.minRadiusPercent = 0.5;
+
+    return custom;
+};
+
+this.StarSpreader = function (helper, custom) {
+
+    if (custom === null) {
+        custom = StarSpreaderCustomization();
+    }
+
+    helper.setBaseValue(custom.spreaderPercent, custom);
+
+    r = helper.wheelRadius * custom.spreaderPercent;
+    rbase = r * custom.minRadiusPercent;
+
+    r = helper.sliceRadius;
+
+    spreaderPathString = [];
+
+    sliceAngle = helper.sliceAngle / helper.navItemCount;
+
+    spreaderPathString.push(helper.MoveTo(helper.startAngle + (helper.navAngle + sliceAngle / 2), rbase));
+    
+    for (var i = 0; i < helper.navItemCount; i++) {
+        startAngle = i * sliceAngle + (helper.navAngle + sliceAngle / 2);
+        middleAngle = startAngle + (sliceAngle / 2);
+        endAngle = startAngle + sliceAngle;
+
+        spreaderPathString.push(helper.LineTo(startAngle, rbase));
+        spreaderPathString.push(helper.LineTo(middleAngle, r));
+        spreaderPathString.push(helper.LineTo(endAngle, rbase));
+    }
+
+    spreaderPathString.push(helper.Close());
+
+    return {
+        spreaderPathString: spreaderPathString,
+        titlePosX: helper.titlePosX,
+        titlePosY: helper.titlePosY
+    };
+};
+
+
+///#source 1 1 /js/source/spreader/wheelnav.spreaderPathEnd.js
+
+    return this;
+};
+
+
+