+++ js/source/spreader/wheelnav.spreaderPath.Pie.js
@@ -1,0 +1,33 @@
+﻿
+this.PieSpreaderCustomization = function () {
+
+    var custom = new spreaderPathCustomization();
+    custom.spreaderRadius = 25;
+    custom.arcBaseRadiusPercent = 1;
+    custom.arcRadiusPercent = 1;
+    custom.startRadiusPercent = 0;
+    return custom;
+};
+
+this.PieSpreader = function (helper, custom) {
+
+    if (custom === null) {
+        custom = PieSpreaderCustomization();
+    }
+
+    helper.setBaseValue(custom.spreaderPercent, custom);
+
+    var arcBaseRadius = helper.sliceRadius * custom.arcBaseRadiusPercent;
+    var arcRadius = helper.sliceRadius * custom.arcRadiusPercent;
+
+    spreaderPathString = [helper.MoveTo(helper.startAngle, arcBaseRadius),
+                 helper.ArcTo(arcRadius, helper.middleAngle, arcBaseRadius),
+                 helper.ArcTo(arcRadius, helper.endAngle, arcBaseRadius),
+                 helper.Close()];
+    
+    return {
+        spreaderPathString: spreaderPathString,
+        titlePosX: helper.titlePosX,
+        titlePosY: helper.titlePosY
+    };
+};