+++ js/source/spreader/wheelnav.spreaderPath.Star.js
@@ -1,0 +1,47 @@
+﻿
+this.StarSpreaderCustomization = function () {
+
+    var custom = new spreaderPathCustomization();
+    custom.minRadiusPercent = 0.5;
+
+    return custom;
+};
+
+this.StarSpreader = function (helper, custom) {
+
+    if (custom === null) {
+        custom = StarSpreaderCustomization();
+    }
+
+    helper.setBaseValue(custom.spreaderPercent, custom);
+
+    r = helper.wheelRadius * custom.spreaderPercent;
+    rbase = r * custom.minRadiusPercent;
+
+    r = helper.sliceRadius;
+
+    spreaderPathString = [];
+
+    sliceAngle = helper.sliceAngle / helper.navItemCount;
+
+    spreaderPathString.push(helper.MoveTo(helper.startAngle + (helper.navAngle + sliceAngle / 2), rbase));
+    
+    for (var i = 0; i < helper.navItemCount; i++) {
+        startAngle = i * sliceAngle + (helper.navAngle + sliceAngle / 2);
+        middleAngle = startAngle + (sliceAngle / 2);
+        endAngle = startAngle + sliceAngle;
+
+        spreaderPathString.push(helper.LineTo(startAngle, rbase));
+        spreaderPathString.push(helper.LineTo(middleAngle, r));
+        spreaderPathString.push(helper.LineTo(endAngle, rbase));
+    }
+
+    spreaderPathString.push(helper.Close());
+
+    return {
+        spreaderPathString: spreaderPathString,
+        titlePosX: helper.titlePosX,
+        titlePosY: helper.titlePosY
+    };
+};
+