+++ js/source/spreader/wheelnav.spreaderPathStart.js
@@ -1,0 +1,18 @@
+﻿/* ======================================================================================= */
+/* Spreader path definitions.                                                              */
+/* ======================================================================================= */
+
+spreaderPath = function () {
+
+    this.NullSpreader = function (helper, custom) {
+
+        helper.setBaseValue(custom);
+
+        return {
+            slicePathString: "",
+            titlePosX: helper.titlePosX,
+            titlePosY: helper.titlePosY
+        };
+    };
+
+